refactor(search): extract ExpressionEvaluationRepository
